Update versions db on self update
