+++ src/bin/juliaup.rs
@@ -23,7 +23,6 @@ use juliaup::{
     command_config_startupselfupdate::run_command_config_startupselfupdate,
     command_config_modifypath::run_command_config_modifypath,
 };
-#[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
 use juliaup::command_selfupdate::run_command_selfupdate;
 use log::info;
 
@@ -81,7 +80,6 @@ enum Juliaup {
     },
     #[clap(name = "info", hide = true)]
     Info { },
-    #[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
     #[clap(subcommand, name = "self")]
     SelfSubCmd(SelfSubCmd),
     // This is used for the cron jobs that we create. By using this UUID for the command
@@ -91,12 +89,14 @@ enum Juliaup {
     SecretSelfUpdate {},
 }
 
-#[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
 #[derive(Parser)]
 /// Manage this juliaup installation
 enum SelfSubCmd {
-    #[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
-    /// Update juliaup itself
+    #[cfg(not(feature = "selfupdate"))]
+    /// Update the Julia versions database
+    Update {},
+    #[cfg(feature = "selfupdate")]
+    /// Update the Julia versions database and juliaup itself
     Update {},
     #[cfg(feature = "selfupdate")]
     /// Configure the channel to use for juliaup updates
@@ -182,9 +182,7 @@ fn main() -> Result<()> {
         Juliaup::Info {} => run_command_info(&paths),
         #[cfg(feature = "selfupdate")]
         Juliaup::SecretSelfUpdate {} => run_command_selfupdate(&paths),
-        #[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
         Juliaup::SelfSubCmd(subcmd) => match subcmd {
-            #[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
             SelfSubCmd::Update {} => run_command_selfupdate(&paths),
             #[cfg(feature = "selfupdate")]
             SelfSubCmd::Channel {channel}  =>  run_command_selfchannel(channel, &paths),
