+++ src/command_selfupdate.rs
@@ -1,5 +1,6 @@
-#[cfg(any(feature = "selfupdate", feature = "windowsstore", feature = "windowsappinstaller"))]
-use anyhow::Result;
+use anyhow::{Result, Context};
+use crate::global_paths::GlobalPaths;
+use crate::operations::update_version_db;
 
 #[cfg(feature = "selfupdate")]
 pub fn run_command_selfupdate(paths: &crate::global_paths::GlobalPaths) -> Result<()> {
@@ -9,6 +10,9 @@ pub fn run_command_selfupdate(paths: &crate::global_paths::GlobalPaths) -> Resul
     use anyhow::Context;
     use crate::operations::{download_juliaup_version,download_extract_sans_parent};
     use crate::{get_juliaup_target, get_own_version};
+
+    update_version_db(paths)
+        .with_context(|| "Failed to uopdate versions db.")?;
 
     let mut config_file =
         load_mut_config_db(paths).with_context(|| "`selfupdate` command failed to load configuration db.")?;
@@ -74,6 +78,9 @@ pub fn run_command_selfupdate(_paths: &crate::global_paths::GlobalPaths) -> Resu
     use anyhow::Context;
     use windows::{core::Interface,Win32::{System::Console::GetConsoleWindow, UI::Shell::IInitializeWithWindow}};
 
+    update_version_db(paths)
+        .with_context(|| "Failed to uopdate versions db.")?;
+
     let update_manager = windows::Services::Store::StoreContext::GetDefault()    
         .with_context(|| "Failed to get the store context.")?;
 
@@ -108,8 +115,9 @@ pub fn run_command_selfupdate(_paths: &crate::global_paths::GlobalPaths) -> Resu
     Ok(())
 }
 
-#[cfg(feature = "windowsappinstaller")]
-pub fn run_command_selfupdate(_paths: &crate::global_paths::GlobalPaths) -> Result<()> {
-    println!("Not yet implemented.");
+#[cfg(not(any(feature = "windowsstore", feature = "selfupdate")))]
+pub fn run_command_selfupdate(paths: &GlobalPaths) -> Result<()> {
+    update_version_db(paths)
+        .with_context(|| "Failed to uopdate versions db.")?;
     Ok(())
 }
